// projects/StarWarAPI/src/app/models/films.model.ts
export interface films {
  count: number;
  next: null;
  previous: null;
  results: filmsRes[];
}

export interface filmsRes {
  title: string;
  episode_id: number;
  opening_crawl: string;
  director: string;
  producer: string;
  release_date: string;
  characters: string[];
  planets: string[];
  starships: string[];
  vehicles: string[];
  species: string[];
  created: string;
  edited: string;
  url: string;
}